+++ ci/docs/build.sh
@@ -23,10 +23,10 @@ cd $PROJECT_WORKSPACE
 export GIT_DESCRIBE_TAG=`git describe --tags`
 export MINOR_VERSION=`echo $GIT_DESCRIBE_TAG | grep -o -E '([0-9]+\.[0-9]+)'`
 
-logger "Check environment..."
+gpuci_logger "Check environment"
 env
 
-logger "Check GPU usage..."
+gpuci_logger "Check GPU usage"
 nvidia-smi
 
 logger "Activate conda env..."
@@ -38,15 +38,21 @@ pip install mockito
 pip install "git+https://github.com/slashnext/SlashNext-URL-Analysis-and-Enrichment.git#egg=slashnext-phishing-ir&subdirectory=Python SDK/src"
 pip install cupy-cuda${CUDA_SHORT}
 
-logger "Check versions..."
+gpuci_logger "Check versions"
 python --version
-conda list
+$CC --version
+$CXX --version
+
+gpuci_logger "Show conda info"
+conda info
+conda config --show-sources
+conda list --show-channel-urls
 
 #clx source build
 ${PROJECT_WORKSPACE}/build.sh clx
 
 #clx Sphinx Build
-logger "Build clx docs..."
+gpuci_logger "Build clx docs"
 cd ${PROJECT_WORKSPACE}/docs
 make html